backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
sqlalchemy==2.0.35
pymysql==1.1.1
cryptography==43.0.1
